--- z-proxy/src/utils.rs
@@ -1,3 +1,0 @@
-// zingo_utils.rs [lib]
-// use: zingo-proxy utils
-//
