+++ zingo-rpc/Cargo.toml
@@ -2,3 +2,16 @@
 name = "zingo-rpc"
 version = "0.1.0"
 edition = "2021"
+
+
+[dependencies]
+http = "0.2.4"
+tokio = { version = "1", features = ["full"] }
+tonic = "0.10.2"
+prost = "0.12"
+bytes = "1.1"
+
+# zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", tag = "zingo_rc.2.1", features = ["lightwalletd-tonic"] }
+zcash_client_backend = { git = "https://github.com/zingolabs/librustzcash.git", branch = "nym_integration", features = ["lightwalletd-tonic"] }
+# zingo-netutils = { git = "https://github.com/aloearev/zingolib.git", branch = "cratify_netutils" }
+zingo-netutils = { git = "https://github.com/zingolabs/zingolib.git", branch = "nym_integration" }
