+++ zingo-rpc/src/lib.rs
@@ -1,0 +1,5 @@
+//! ZCash's lightwallet service and darkside_service RPC implementations.
+
+pub mod primitives;
+pub mod service;
+mod utils;
